Move id logic to extension method
